+++ packages/itwin/property-grid/src/e2e-tests/PropertyGrid.test.ts-snapshots/should-match-image-snapshot-stage-panel-size---narrow-single-element-selected-1-chromium-win32.png
