+++ packages/itwin/tree-widget/src/e2e-tests/TreeWidget.test.ts-snapshots/should-match-image-snapshot-expanded-tree-node-1-chromium-win32.png
