+++ packages/itwin/tree-widget/src/e2e-tests/TreeWidget.test.ts-snapshots/should-match-image-snapshot-initial-tree-1-chromium-win32.png
