+++ packages/itwin/tree-widget/src/e2e-tests/TreeWidget.test.ts-snapshots/should-match-image-snapshot-search-1-chromium-win32.png
